User Action Accept/Declined and new Request tested
