+++ antd-mobile/src/components/users/profilesections/settings/SettingsDialog.jsx
@@ -16,9 +16,10 @@ import {
 import { useNavigate } from "react-router-dom";
 import "../../../../styles/settings.css";
 import MobileImageUploader from "../../../common/MobileImageUploader";
-import { getUserById,updateUserData } from "../../../../services/api";
+//import { getUserById,updateUserData } from "../../../../services/api";
 import { AuthContext } from "../../../../context/AuthContext";
 import Preferences from "./Prefrences";
+import {getUserById, updateUserData } from "../../../../services/api";
 
 export const SettingsDialog = () => {
   const {user} = useContext(AuthContext)
